+++ package.json
@@ -7,6 +7,7 @@
     "test": "mocha --timeout 10000 --recursive",
     "endpoint-tests": "mocha --timeout 10000 test/endpoint-tests/*.test.js",
     "unit-tests": "mocha test/unit-tests/*.test.js",
+    "single-test": "mocha test/single-test/*.test.js",
     "start": "SET NODE_ENV=dev && node server.js",
     "restart": "SET NODE_ENV=dev && nodemon server.js",
     "build": "webpack",
